Move useEmojiKeyboard
